+++ src/main/kotlin/icu/windea/pls/core/expression/ParadoxDataExpression.kt
@@ -2,6 +2,7 @@ package icu.windea.pls.core.expression
 
 import com.intellij.openapi.progress.*
 import icu.windea.pls.*
+import icu.windea.pls.config.config.*
 import icu.windea.pls.core.*
 import icu.windea.pls.core.expression.ParadoxDataExpression.*
 import icu.windea.pls.script.psi.*
@@ -33,11 +34,14 @@ object UnknownParadoxDataExpression: AbstractExpression(PlsConstants.unknownStri
 	override val isKey: Boolean = false
 }
 
-fun Resolver.resolve(element: ParadoxScriptExpressionElement): ParadoxDataExpression {
+fun Resolver.resolve(element: ParadoxScriptExpressionElement, matchType: Int = CwtConfigMatchType.DEFAULT): ParadoxDataExpression {
 	return when {
 		element is ParadoxScriptScriptedVariableReference -> {
 			ProgressManager.checkCanceled() //这是必要的
-			val valueElement = element.referenceValue ?: return UnknownParadoxDataExpression
+			val valueElement = when {
+				matchType == CwtConfigMatchType.STATIC -> return UnknownParadoxDataExpression 
+				else -> element.referenceValue ?: return UnknownParadoxDataExpression
+			}
 			ParadoxDataExpressionImpl(valueElement.value, valueElement.type, valueElement.text.isLeftQuoted(), false)
 		}
 		element.type == ParadoxDataType.BlockType -> {