+++ src/main/kotlin/icu/windea/pls/core/index/ParadoxValueSetValueIndex.kt
@@ -3,11 +3,11 @@ package icu.windea.pls.core.index
 import com.intellij.psi.stubs.*
 import icu.windea.pls.script.psi.*
 
-class ParadoxValueSetValueIndex : StringStubIndexExtension<ParadoxScriptString>() {
+class ParadoxValueSetValueIndex : StringStubIndexExtension<ParadoxScriptStringExpressionElement>() {
     companion object {
-        @JvmField val KEY = StubIndexKey.createIndexKey<String, ParadoxScriptString>("paradox.valueSetValue.index")
+        @JvmField val KEY = StubIndexKey.createIndexKey<String, ParadoxScriptStringExpressionElement>("paradox.valueSetValue.index")
         private const val VERSION = 17 //0.9.2
-        private const val CACHE_SIZE = 2048
+        private const val CACHE_SIZE = 10 * 1024 //5675 in stellaris@3.6
     }
     
     override fun getKey() = KEY